+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Recyclarr will now continue if `git fetch` fails for any repos, so long as there is an existing,
+  valid clone to use.
+
 ## [5.0.0] - 2023-06-22
 
 This release contains **BREAKING CHANGES**. See the [v5.0 Upgrade Guide][breaking5] for required